full documentation, empty test files, small refactors, clover added, pom modifications

+++ src/main/java/unicrush/model/Game.java
@@ -80,7 +80,7 @@ public interface Game {
      *
      * @return the current score of a player
      */
-    public long getPlayerScore();
+    public int getPlayerScore();
     
     /**
      * Used for modifying the score of a player.
